chore(frontend): retire legacy Agenta 1.0 onboarding
Remove the pre-pivot onboarding surfaces the new playground-native flow
replaces: the /get-started three-option page (OSS + EE), the home
WelcomeCards banner and the now-orphaned old app-management home page,
and the EE demo-applications section.

Park the reusable pieces behind kill switches instead of deleting them:
hide the sidebar Get Started Guide item, stop mounting the onboarding
checklist widget (and disable its new-user auto-open), and gate the
nextstepjs guided-tour engine behind ONBOARDING_TOURS_ENABLED so it never
runs. Route new users straight to the resolved post-login path instead of
/get-started.
