+++ web/oss/src/components/Layout/Layout.tsx
@@ -72,7 +72,6 @@ const layoutRouteFlagsAtom = atom<LayoutRouteFlags>((get) => {
         isAuthRoute:
             pathname.includes("/auth") ||
             pathname.includes("/post-signup") ||
-            pathname.includes("/get-started") ||
             pathname.startsWith("/workspaces/accept"),
         isAppRoute: routeLayer === "app",
         isPlayground: pathname.includes("/playground"),
