+++ web/oss/src/components/Onboarding/index.ts
@@ -1,6 +1,10 @@
 export {default as OnboardingCard} from "./OnboardingCard"
 export {default as OnboardingProvider} from "./OnboardingProvider"
-export {OnboardingWidget} from "./Widget"
+// OnboardingWidget is mounted directly (behind a dynamic import) from AppGlobalWrappers,
+// not re-exported here — that mount is currently disabled. Re-exporting it from this
+// barrel would pull its top-level `@agentaai/nextstepjs` import into every module that
+// imports anything else from this barrel (e.g. useOnboardingTour), defeating the
+// tours-disabled kill switch. Import it directly from "./Widget" if ever needed again.
 export {useOnboardingTour} from "./hooks/useOnboardingTour"
 
 // Tours
