--- web/oss/src/components/pages/app-management/components/WelcomeCardsSection/assets/store/welcomeCards.ts
@@ -1,5 +1,0 @@
-import {atomWithStorage} from "jotai/utils"
-
-const STORAGE_KEY = "agenta:app-management:welcome-cards-dismissed"
-
-export const welcomeCardsDismissedAtom = atomWithStorage<boolean>(STORAGE_KEY, false)
